Unclutters the Application class, finally
